Split concepts out into self contained files in feedback editor
